+++ src/BlockFermionicUniTensor.cpp
@@ -1808,7 +1808,7 @@ namespace cytnx {
     //[21 Aug 2024] This is a copy from BlockUniTensor;
     Scalar out(0, this->dtype());
     for (auto &block : this->_blocks) {
-      double bn = linalg::norm(block);
+      double bn = double(linalg::norm(block));
       out += Scalar(bn * bn);
     }
     out = sqrt(out);
@@ -1948,10 +1948,10 @@ namespace cytnx {
     //[21 Aug 2024] This is a copy from BlockUniTensor;
     Scalar t;
     if (this->_blocks.size()) {
-      double n0 = linalg::norm(this->_blocks[0]);
+      double n0 = double(linalg::norm(this->_blocks[0]));
       t = Scalar(n0 * n0);
       for (int blk = 1; blk < this->_blocks.size(); blk++) {
-        double nblk = linalg::norm(this->_blocks[blk]);
+        double nblk = double(linalg::norm(this->_blocks[blk]));
         t += Scalar(nblk * nblk);
       }
